separate hierCluster renderer and interactivity code from data processing code
